+++ bial/inc/FastIncreasingFifoBucketQueue.hpp
@@ -75,7 +75,7 @@ namespace Bial {
      * @brief Removes the next element in the bucket queue. 
      * @warning none. 
      */
-    int Remove( );
+    size_t Remove( );
 
   protected:
