+++ bial/inc/InfBucketQueue.hpp
@@ -100,7 +100,7 @@ namespace Bial {
      * @brief Removes the next element in the bucket queue. 
      * @warning none. 
      */
-    int Remove( );
+    size_t Remove( );
 
   protected:
