+++ bial/src/FastBucketQueue.cpp
@@ -125,7 +125,7 @@ namespace Bial {
     }
   }
 
-  int FastBucketQueue::Remove( ) {
+  size_t FastBucketQueue::Remove( ) {
     try {
       COMMENT( "Current amount of elements prior to removal: " << this->elements, 3 );
       COMMENT( "Check if queue is initialized.", 4 );
